+++ test/functional/rpc_rawtransaction.py
@@ -620,10 +620,10 @@ class RawTransactionsTest(BitcoinTestFramework):
         testres = self.nodes[2].testmempoolaccept(
             [rawTxSigned['hex']], 500.00)[0]
         assert_equal(testres['allowed'], False)
-        assert_equal(testres['reject-reason'], 'absurdly-high-fee')
+        assert_equal(testres['reject-reason'], 'max-fee-exceeded')
         # and sendrawtransaction should throw
-        assert_raises_rpc_error(-26,
-                                "absurdly-high-fee",
+        assert_raises_rpc_error(-25,
+                                'Fee exceeds maximum configured by -maxtxfee',
                                 self.nodes[2].sendrawtransaction,
                                 rawTxSigned['hex'],
                                 10.00)
@@ -652,10 +652,10 @@ class RawTransactionsTest(BitcoinTestFramework):
         # Thus, testmempoolaccept should reject
         testres = self.nodes[2].testmempoolaccept([rawTxSigned['hex']])[0]
         assert_equal(testres['allowed'], False)
-        assert_equal(testres['reject-reason'], 'absurdly-high-fee')
+        assert_equal(testres['reject-reason'], 'max-fee-exceeded')
         # and sendrawtransaction should throw
-        assert_raises_rpc_error(-26,
-                                "absurdly-high-fee",
+        assert_raises_rpc_error(-25,
+                                'Fee exceeds maximum configured by -maxtxfee',
                                 self.nodes[2].sendrawtransaction,
                                 rawTxSigned['hex'])
         # and the following calls should both succeed